requirements.txt
streamlit>=1.32.0
opencv-python>=4.9.0
Pillow>=10.0.0
numpy>=1.24.0
bcrypt>=4.1.0
PyJWT>=2.8.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
email-validator>=2.1.0